fix: 같은 Crew가 여러명이 들어가는 오류 수정

+++ src/main/java/attendance/model/AttendanceWarning.java
@@ -15,7 +15,7 @@ public enum AttendanceWarning {
     }
 
 
-    public static AttendanceWarning from(int absenceCount) {
+    public static AttendanceWarning from(long absenceCount) {
         if (absenceCount > 5) {
             return 제적;
         }
